+++ src/Containers/OhmmsSoA/tests/CMakeLists.txt
@@ -19,7 +19,7 @@ SET(UTEST_NAME deterministic-unit_test_${SRC_DIR})
 
 
 ADD_EXECUTABLE(${UTEST_EXE} test_vector_soa.cpp)
-TARGET_LINK_LIBRARIES(${UTEST_EXE} catch_main qmcutil)
+TARGET_LINK_LIBRARIES(${UTEST_EXE} catch_main containers)
 
 ADD_UNIT_TEST(${UTEST_NAME} "${QMCPACK_UNIT_TEST_DIR}/${UTEST_EXE}")
