+++ src/type_traits/tests/CMakeLists.txt
@@ -24,6 +24,6 @@ if(QMC_CUDA)
   SET(TEST_SRCS test_cuda_types.cpp)
 endif(QMC_CUDA)
 ADD_EXECUTABLE(${UTEST_EXE} ${TEST_SRCS})
-TARGET_LINK_LIBRARIES(${UTEST_EXE} catch_main)
+TARGET_LINK_LIBRARIES(${UTEST_EXE} catch_main containers)
 
 ADD_UNIT_TEST(${UTEST_NAME} "${QMCPACK_UNIT_TEST_DIR}/${UTEST_EXE}")
